update/rename mwixnet onion classes
