+++ api/src/owner.rs
@@ -31,7 +31,7 @@ use crate::libwallet::api_impl::{owner, owner_updater};
 use crate::libwallet::contract::types::{
 	ContractNewArgsAPI, ContractRevokeArgsAPI, ContractSetupArgsAPI,
 };
-use crate::libwallet::mwmixnet::types::{MixnetReqCreationParams, SwapReq};
+use crate::libwallet::mwixnet::types::{MixnetReqCreationParams, SwapReq};
 use crate::libwallet::{
 	AcctPathMapping, BuiltOutput, Error, InitTxArgs, IssueInvoiceTxArgs, NodeClient,
 	NodeHeightResult, OutputCommitMapping, PaymentProof, Slate, Slatepack, SlatepackAddress,
@@ -831,7 +831,7 @@ where
 	}
 
 	/// Create MXMixnet request
-	pub fn create_mwmixnet_req(
+	pub fn create_mwixnet_req(
 		&self,
 		keychain_mask: Option<&SecretKey>,
 		params: &MixnetReqCreationParams,
@@ -840,7 +840,7 @@ where
 	) -> Result<SwapReq, Error> {
 		let mut w_lock = self.wallet_inst.lock();
 		let w = w_lock.lc_provider()?.wallet_inst()?;
-		owner::create_mwmixnet_req(&mut **w, keychain_mask, params, slate)
+		owner::create_mwixnet_req(&mut **w, keychain_mask, params, slate)
 	}
 
 	/// Processes an invoice tranaction created by another party, essentially
