+++ libwallet/src/mwixnet/onion/util.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Util fns for mwmixnet
+//! Util fns for mwixnet
 //! TODO: possibly redundant, check or move elsewhere
 
 use grin_core::ser::{self, Readable, Reader, Writeable, Writer};
@@ -37,7 +37,7 @@ use std::convert::TryInto;
 /// # Example
 ///
 /// ```
-///	use grin_wallet_libwallet::mwmixnet::onion::util::write_optional;
+///	use grin_wallet_libwallet::mwixnet::onion::util::write_optional;
 /// let mut writer:Vec<u8> = vec![];
 /// let optional_value: Option<u32> = Some(10);
 /// //write_optional(&mut writer, &optional_value);
@@ -75,7 +75,7 @@ pub fn write_optional<O: Writeable, W: Writer>(
 /// # Example
 ///
 /// ```
-///	use grin_wallet_libwallet::mwmixnet::onion::util::read_optional;
+///	use grin_wallet_libwallet::mwixnet::onion::util::read_optional;
 /// use grin_core::ser::{BinReader, ProtocolVersion, DeserializationMode};
 /// let mut buf: &[u8] = &[1, 0, 0, 0, 10];
 /// let mut reader = BinReader::new(&mut buf, ProtocolVersion::local(), DeserializationMode::default());
@@ -106,7 +106,7 @@ pub fn read_optional<O: Readable, R: Reader>(reader: &mut R) -> Result<Option<O>
 /// # Example
 ///
 /// ```
-///	use grin_wallet_libwallet::mwmixnet::onion::util::vec_to_array;
+///	use grin_wallet_libwallet::mwixnet::onion::util::vec_to_array;
 /// let v = vec![0, 1, 2, 3, 4, 5];
 /// let a = vec_to_array::<4>(&v).unwrap();
 /// assert_eq!(a, [0, 1, 2, 3]);
